Fix compilation error (#39)
