+++ src/core/util.cpp
@@ -21,3 +21,25 @@ Color operator-(const Color& left, const Color& right)
 #endif
 
 }  // namespacs sf
+
+FloatColor uncompressColor(const sf::Color &color)
+{
+    // color.r = 255.0 * std::pow(image[y * w + x].r / 255.0, 2.2);
+    // color.g = 255.0 * std::pow(image[y * w + x].g / 255.0, 2.2);
+    // color.b = 255.0 * std::pow(image[y * w + x].b / 255.0, 2.2);
+
+    return FloatColor(255.0 * std::pow(color.r / 255.0, 2.2),
+                      255.0 * std::pow(color.g / 255.0, 2.2),
+                      255.0 * std::pow(color.b / 255.0, 2.2));
+}
+
+sf::Color compressColor(const FloatColor &color)
+{
+    // image[y * w + x].r = 255.0 * std::pow(color.r / 255.0, 1 / 2.2);
+    // image[y * w + x].g = 255.0 * std::pow(color.g / 255.0, 1 / 2.2);
+    // image[y * w + x].b = 255.0 * std::pow(color.b / 255.0, 1 / 2.2);
+
+    return sf::Color(255.0 * std::pow(color.r / 255.0, 1 / 2.2),
+                     255.0 * std::pow(color.g / 255.0, 1 / 2.2),
+                     255.0 * std::pow(color.b / 255.0, 1 / 2.2));
+}
